added favicon, auth refactor
