+++ frontend/index.html
@@ -7,6 +7,7 @@
     <link rel="preconnect" href="https://rsms.me/">
     <link rel="stylesheet" href="https://rsms.me/inter/inter.css">
     <link href="/css/output.css" rel="stylesheet">
+    <link rel="icon" type="image/svg" href="/assets/ping-pong.svg">
     <title>ft_transcendence</title>
 </head>
